breaking apart compiler logic
