+++ crates/runmat-vm/src/compiler/core.rs
@@ -1,3 +1,4 @@
+use crate::compiler::end_expr::{end_numeric_expr, expr_is_one, range_dynamic_end_spec};
 use crate::compiler::CompileError;
 use crate::functions::UserFunction;
 use crate::instr::{EmitLabel, EndExpr, Instr};
@@ -54,133 +55,6 @@ struct StochasticEvolutionPlan<'a> {
     drift: &'a HirExpr,
     scale: &'a HirExpr,
     steps: &'a HirExpr,
-}
-
-fn expr_is_one(expr: &HirExpr) -> bool {
-    parse_number(expr)
-        .map(|v| (v - 1.0).abs() < 1e-9)
-        .unwrap_or(false)
-}
-
-fn parse_number(expr: &HirExpr) -> Option<f64> {
-    if let runmat_hir::HirExprKind::Number(raw) = &expr.kind {
-        raw.parse::<f64>().ok()
-    } else {
-        None
-    }
-}
-
-fn end_numeric_expr(expr: &HirExpr) -> Option<EndExpr> {
-    end_numeric_expr_from_expr(&expr.kind)
-}
-
-fn numeric_expr_any(expr: &HirExpr) -> Option<EndExpr> {
-    end_numeric_expr_internal(&expr.kind).map(|(e, _)| e)
-}
-
-fn range_dynamic_end_spec(expr: &HirExpr) -> Option<(Option<EndExpr>, Option<EndExpr>, EndExpr)> {
-    let HirExprKind::Range(start, step, end) = &expr.kind else {
-        return None;
-    };
-    let start_end = end_numeric_expr(start);
-    let step_end = step.as_ref().and_then(|s| end_numeric_expr(s));
-    let end_end = end_numeric_expr(end);
-    if start_end.is_none() && step_end.is_none() && end_end.is_none() {
-        return None;
-    }
-    let resolved_end = if let Some(e) = end_end {
-        e
-    } else {
-        numeric_expr_any(end)?
-    };
-    Some((start_end, step_end, resolved_end))
-}
-
-fn end_numeric_expr_from_expr(kind: &HirExprKind) -> Option<EndExpr> {
-    let (expr, has_end) = end_numeric_expr_internal(kind)?;
-    if has_end {
-        Some(expr)
-    } else {
-        None
-    }
-}
-
-fn end_numeric_expr_internal(kind: &HirExprKind) -> Option<(EndExpr, bool)> {
-    match kind {
-        HirExprKind::End => Some((EndExpr::End, true)),
-        HirExprKind::Number(s) => s.parse::<f64>().ok().map(|v| (EndExpr::Const(v), false)),
-        HirExprKind::Var(id) => Some((EndExpr::Var(id.0), false)),
-        HirExprKind::Unary(op, inner) => {
-            let (child, has_end) = end_numeric_expr_internal(&inner.kind)?;
-            match op {
-                runmat_parser::UnOp::Plus => Some((EndExpr::Pos(Box::new(child)), has_end)),
-                runmat_parser::UnOp::Minus => Some((EndExpr::Neg(Box::new(child)), has_end)),
-                _ => None,
-            }
-        }
-        HirExprKind::FuncCall(name, args) => {
-            let mut out_args = Vec::with_capacity(args.len());
-            let mut has_end = false;
-            for arg in args {
-                let (e, h) = end_numeric_expr_internal(&arg.kind)?;
-                out_args.push(e);
-                has_end |= h;
-            }
-            let lname = name.to_ascii_lowercase();
-            if args.len() == 1 {
-                let single = out_args.into_iter().next().unwrap_or(EndExpr::Const(0.0));
-                let out = match lname.as_str() {
-                    "floor" => EndExpr::Floor(Box::new(single)),
-                    "ceil" | "ceiling" => EndExpr::Ceil(Box::new(single)),
-                    "round" => EndExpr::Round(Box::new(single)),
-                    "fix" => EndExpr::Fix(Box::new(single)),
-                    _ => EndExpr::Call(name.clone(), vec![single]),
-                };
-                Some((out, has_end))
-            } else {
-                Some((EndExpr::Call(name.clone(), out_args), has_end))
-            }
-        }
-        HirExprKind::Binary(left, op, right) => {
-            let (lhs, left_has_end) = end_numeric_expr_internal(&left.kind)?;
-            let (rhs, right_has_end) = end_numeric_expr_internal(&right.kind)?;
-            let has_end = left_has_end || right_has_end;
-            match op {
-                runmat_parser::BinOp::Add => {
-                    Some((EndExpr::Add(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::Sub => {
-                    Some((EndExpr::Sub(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::Mul => {
-                    Some((EndExpr::Mul(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::ElemMul => {
-                    Some((EndExpr::Mul(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::RightDiv => {
-                    Some((EndExpr::Div(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::ElemDiv => {
-                    Some((EndExpr::Div(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::LeftDiv => {
-                    Some((EndExpr::LeftDiv(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::ElemLeftDiv => {
-                    Some((EndExpr::LeftDiv(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::Pow => {
-                    Some((EndExpr::Pow(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                runmat_parser::BinOp::ElemPow => {
-                    Some((EndExpr::Pow(Box::new(lhs), Box::new(rhs)), has_end))
-                }
-                _ => None,
-            }
-        }
-        _ => None,
-    }
 }
 
 fn stochastic_evolution_disabled() -> bool {
@@ -420,152 +294,6 @@ impl Compiler {
         };
         (is_static, acc_str)
     }
-    fn expr_contains_end(expr: &runmat_hir::HirExpr) -> bool {
-        use runmat_hir::HirExprKind as K;
-        match &expr.kind {
-            K::End => true,
-            K::Unary(_, e) => Self::expr_contains_end(e),
-            K::Binary(a, _, b) => Self::expr_contains_end(a) || Self::expr_contains_end(b),
-            K::Range(a, b, c) => {
-                Self::expr_contains_end(a)
-                    || b.as_ref()
-                        .map(|e| Self::expr_contains_end(e))
-                        .unwrap_or(false)
-                    || Self::expr_contains_end(c)
-            }
-            K::FuncCall(_, args) => args.iter().any(Self::expr_contains_end),
-            K::Tensor(rows) | K::Cell(rows) => rows
-                .iter()
-                .flat_map(|r| r.iter())
-                .any(Self::expr_contains_end),
-            K::Index(base, idxs) | K::IndexCell(base, idxs) => {
-                if Self::expr_contains_end(base) {
-                    return true;
-                }
-                idxs.iter().any(Self::expr_contains_end)
-            }
-            _ => false,
-        }
-    }
-
-    fn allow_implicit_struct_materialization(ty: &Type) -> bool {
-        matches!(ty, Type::Unknown | Type::Struct { .. })
-    }
-
-    fn emit_store_back_member_chain(
-        &mut self,
-        base: &runmat_hir::HirExpr,
-    ) -> Result<(), CompileError> {
-        match &base.kind {
-            runmat_hir::HirExprKind::Var(var_id) => {
-                self.emit(Instr::StoreVar(var_id.0));
-                Ok(())
-            }
-            runmat_hir::HirExprKind::Member(parent, field) => {
-                self.compile_member_base_for_assignment(parent)?;
-                self.emit(Instr::Swap);
-                self.emit(Instr::StoreMemberOrInit(field.clone()));
-                self.emit_store_back_member_chain(parent)
-            }
-            runmat_hir::HirExprKind::MemberDynamic(parent, name_expr) => {
-                let tmp = self.alloc_temp();
-                self.emit(Instr::StoreVar(tmp));
-                self.compile_member_base_for_assignment(parent)?;
-                self.compile_expr(name_expr)?;
-                self.emit(Instr::LoadVar(tmp));
-                self.emit(Instr::StoreMemberDynamicOrInit);
-                self.emit_store_back_member_chain(parent)
-            }
-            _ => Ok(()),
-        }
-    }
-
-    fn compile_member_base_for_assignment(
-        &mut self,
-        base: &runmat_hir::HirExpr,
-    ) -> Result<(), CompileError> {
-        match &base.kind {
-            runmat_hir::HirExprKind::Member(parent, field) => {
-                self.compile_member_base_for_assignment(parent)?;
-                self.emit(Instr::LoadMemberOrInit(field.clone()));
-                Ok(())
-            }
-            runmat_hir::HirExprKind::MemberDynamic(parent, name_expr) => {
-                self.compile_member_base_for_assignment(parent)?;
-                self.compile_expr(name_expr)?;
-                self.emit(Instr::LoadMemberDynamicOrInit);
-                Ok(())
-            }
-            _ => self.compile_expr(base),
-        }
-    }
-    #[allow(clippy::only_used_in_recursion)]
-    fn collect_free_vars(
-        &self,
-        expr: &runmat_hir::HirExpr,
-        bound: &std::collections::HashSet<runmat_hir::VarId>,
-        seen: &mut std::collections::HashSet<runmat_hir::VarId>,
-        out: &mut Vec<runmat_hir::VarId>,
-    ) {
-        use runmat_hir::HirExprKind as K;
-        match &expr.kind {
-            K::Var(id) => {
-                if !bound.contains(id) && !seen.contains(id) {
-                    seen.insert(*id);
-                    out.push(*id);
-                }
-            }
-            K::Unary(_, e) => self.collect_free_vars(e, bound, seen, out),
-            K::Binary(a, _, b) => {
-                self.collect_free_vars(a, bound, seen, out);
-                self.collect_free_vars(b, bound, seen, out);
-            }
-            K::Tensor(rows) | K::Cell(rows) => {
-                for row in rows {
-                    for e in row {
-                        self.collect_free_vars(e, bound, seen, out);
-                    }
-                }
-            }
-            K::Index(base, idxs) | K::IndexCell(base, idxs) => {
-                self.collect_free_vars(base, bound, seen, out);
-                for i in idxs {
-                    self.collect_free_vars(i, bound, seen, out);
-                }
-            }
-            K::Range(s, st, e) => {
-                self.collect_free_vars(s, bound, seen, out);
-                if let Some(st) = st {
-                    self.collect_free_vars(st, bound, seen, out);
-                }
-                self.collect_free_vars(e, bound, seen, out);
-            }
-            K::FuncCall(_, args) => {
-                for a in args {
-                    self.collect_free_vars(a, bound, seen, out);
-                }
-            }
-            K::MethodCall(base, _, args) | K::DottedInvoke(base, _, args) => {
-                self.collect_free_vars(base, bound, seen, out);
-                for a in args {
-                    self.collect_free_vars(a, bound, seen, out);
-                }
-            }
-            K::Member(base, _) => self.collect_free_vars(base, bound, seen, out),
-            K::MemberDynamic(base, name) => {
-                self.collect_free_vars(base, bound, seen, out);
-                self.collect_free_vars(name, bound, seen, out);
-            }
-            K::AnonFunc { params, body } => {
-                let mut new_bound = bound.clone();
-                for p in params {
-                    new_bound.insert(*p);
-                }
-                self.collect_free_vars(body, &new_bound, seen, out);
-            }
-            _ => {}
-        }
-    }
     pub fn new(prog: &HirProgram) -> Self {
         let mut max_var = 0;
         fn visit_expr(expr: &HirExpr, max: &mut usize) {
@@ -750,7 +478,7 @@ impl Compiler {
         }
     }
 
-    fn alloc_temp(&mut self) -> usize {
+    pub(crate) fn alloc_temp(&mut self) -> usize {
         let id = self.var_count;
         self.var_count += 1;
         if self.var_types.len() <= id {
@@ -786,7 +514,7 @@ impl Compiler {
         self.instructions[idx] = instr;
     }
 
-    fn compile_error(&self, message: impl Into<String>) -> CompileError {
+    pub(crate) fn compile_error(&self, message: impl Into<String>) -> CompileError {
         let mut err = CompileError::new(message);
         if let Some(span) = self.current_span {
             err = err.with_span(span);
@@ -1025,187 +753,14 @@ impl Compiler {
                 has_varargin,
                 has_varargout,
                 ..
-            } => {
-                let mut max_local_var = 0;
-                fn visit_expr_for_vars(expr: &HirExpr, max: &mut usize) {
-                    match &expr.kind {
-                        HirExprKind::Var(id) => {
-                            if id.0 + 1 > *max {
-                                *max = id.0 + 1;
-                            }
-                        }
-                        HirExprKind::Unary(_, e) => visit_expr_for_vars(e, max),
-                        HirExprKind::Binary(a, _, b) => {
-                            visit_expr_for_vars(a, max);
-                            visit_expr_for_vars(b, max);
-                        }
-                        HirExprKind::Tensor(rows) | HirExprKind::Cell(rows) => {
-                            for row in rows {
-                                for elem in row {
-                                    visit_expr_for_vars(elem, max);
-                                }
-                            }
-                        }
-                        HirExprKind::Index(base, indices)
-                        | HirExprKind::IndexCell(base, indices) => {
-                            visit_expr_for_vars(base, max);
-                            for idx in indices {
-                                visit_expr_for_vars(idx, max);
-                            }
-                        }
-                        HirExprKind::Range(start, step, end) => {
-                            visit_expr_for_vars(start, max);
-                            if let Some(step) = step {
-                                visit_expr_for_vars(step, max);
-                            }
-                            visit_expr_for_vars(end, max);
-                        }
-                        HirExprKind::FuncCall(_, args) | HirExprKind::MethodCall(_, _, args) => {
-                            for arg in args {
-                                visit_expr_for_vars(arg, max);
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-                fn visit_stmt_for_vars(stmt: &HirStmt, max: &mut usize) {
-                    match stmt {
-                        HirStmt::ExprStmt(expr, _, _) => visit_expr_for_vars(expr, max),
-                        HirStmt::Assign(id, expr, _, _) => {
-                            if id.0 + 1 > *max {
-                                *max = id.0 + 1;
-                            }
-                            visit_expr_for_vars(expr, max);
-                        }
-                        HirStmt::If {
-                            cond,
-                            then_body,
-                            elseif_blocks,
-                            else_body,
-                            ..
-                        } => {
-                            visit_expr_for_vars(cond, max);
-                            for stmt in then_body {
-                                visit_stmt_for_vars(stmt, max);
-                            }
-                            for (cond, body) in elseif_blocks {
-                                visit_expr_for_vars(cond, max);
-                                for stmt in body {
-                                    visit_stmt_for_vars(stmt, max);
-                                }
-                            }
-                            if let Some(body) = else_body {
-                                for stmt in body {
-                                    visit_stmt_for_vars(stmt, max);
-                                }
-                            }
-                        }
-                        HirStmt::While { cond, body, .. } => {
-                            visit_expr_for_vars(cond, max);
-                            for stmt in body {
-                                visit_stmt_for_vars(stmt, max);
-                            }
-                        }
-                        HirStmt::For {
-                            var, expr, body, ..
-                        } => {
-                            if var.0 + 1 > *max {
-                                *max = var.0 + 1;
-                            }
-                            visit_expr_for_vars(expr, max);
-                            for stmt in body {
-                                visit_stmt_for_vars(stmt, max);
-                            }
-                        }
-                        HirStmt::Break(_) | HirStmt::Continue(_) | HirStmt::Return(_) => {}
-                        HirStmt::Switch {
-                            expr,
-                            cases,
-                            otherwise,
-                            ..
-                        } => {
-                            visit_expr_for_vars(expr, max);
-                            for (c, b) in cases {
-                                visit_expr_for_vars(c, max);
-                                for s in b {
-                                    visit_stmt_for_vars(s, max);
-                                }
-                            }
-                            if let Some(b) = otherwise {
-                                for s in b {
-                                    visit_stmt_for_vars(s, max);
-                                }
-                            }
-                        }
-                        HirStmt::TryCatch {
-                            try_body,
-                            catch_var,
-                            catch_body,
-                            ..
-                        } => {
-                            if let Some(v) = catch_var {
-                                if v.0 + 1 > *max {
-                                    *max = v.0 + 1;
-                                }
-                            }
-                            for s in try_body {
-                                visit_stmt_for_vars(s, max);
-                            }
-                            for s in catch_body {
-                                visit_stmt_for_vars(s, max);
-                            }
-                        }
-                        HirStmt::Global(vars, _) | HirStmt::Persistent(vars, _) => {
-                            for (v, _name) in vars {
-                                if v.0 + 1 > *max {
-                                    *max = v.0 + 1;
-                                }
-                            }
-                        }
-                        HirStmt::AssignLValue(_, expr, _, _) => visit_expr_for_vars(expr, max),
-                        HirStmt::MultiAssign(vars, expr, _, _) => {
-                            for v in vars.iter().flatten() {
-                                if v.0 + 1 > *max {
-                                    *max = v.0 + 1;
-                                }
-                            }
-                            visit_expr_for_vars(expr, max);
-                        }
-                        HirStmt::Function { .. }
-                        | HirStmt::ClassDef { .. }
-                        | HirStmt::Import { .. } => {}
-                    }
-                }
-                for stmt in body {
-                    visit_stmt_for_vars(stmt, &mut max_local_var);
-                }
-                let var_map =
-                    runmat_hir::remapping::create_complete_function_var_map(params, outputs, body);
-                let local_var_count = var_map.len();
-                if local_var_count > max_local_var {
-                    max_local_var = local_var_count;
-                }
-                let mut func_var_types = vec![Type::Unknown; local_var_count];
-                for (orig, local) in &var_map {
-                    if let Some(ty) = self.var_types.get(orig.0) {
-                        if let Some(slot) = func_var_types.get_mut(local.0) {
-                            *slot = ty.clone();
-                        }
-                    }
-                }
-                let user_func = UserFunction {
-                    name: name.clone(),
-                    params: params.clone(),
-                    outputs: outputs.clone(),
-                    body: body.clone(),
-                    local_var_count: max_local_var,
-                    has_varargin: *has_varargin,
-                    has_varargout: *has_varargout,
-                    var_types: func_var_types,
-                    source_id: None,
-                };
-                self.functions.insert(name.clone(), user_func);
-            }
+            } => self.compile_function_stmt(
+                name,
+                params,
+                outputs,
+                body,
+                *has_varargin,
+                *has_varargout,
+            )?,
             HirStmt::Switch {
                 expr,
                 cases,
@@ -1271,489 +826,7 @@ impl Compiler {
                 let end_pc = self.instructions.len();
                 self.patch(jmp_end, Instr::Jump(end_pc));
             }
-            HirStmt::AssignLValue(lv, rhs, _, _) => {
-                match lv {
-                    runmat_hir::HirLValue::Index(base, indices) => {
-                        if let runmat_hir::HirExprKind::Var(var_id) = base.kind {
-                            // Load base variable first
-                            self.emit(Instr::LoadVar(var_id.0));
-                            // Compute masks and numeric indices as in IndexSlice
-                            let has_colon = indices
-                                .iter()
-                                .any(|e| matches!(e.kind, runmat_hir::HirExprKind::Colon));
-                            let has_end = indices
-                                .iter()
-                                .any(|e| matches!(e.kind, runmat_hir::HirExprKind::End));
-                            let has_vector = indices.iter().any(|e| {
-                                matches!(
-                                    e.kind,
-                                    HirExprKind::Range(_, _, _) | HirExprKind::Tensor(_)
-                                ) || matches!(
-                                    e.ty,
-                                    runmat_hir::Type::Tensor { .. }
-                                        | runmat_hir::Type::Bool
-                                        | runmat_hir::Type::Logical { .. }
-                                )
-                            });
-                            if has_colon || has_end || has_vector || indices.len() > 2 {
-                                let mut colon_mask: u32 = 0;
-                                let mut end_mask: u32 = 0;
-                                let mut numeric_count = 0usize;
-                                let mut end_offsets: Vec<(usize, EndExpr)> = Vec::new();
-                                for (dim, index) in indices.iter().enumerate() {
-                                    if matches!(index.kind, runmat_hir::HirExprKind::Colon) {
-                                        colon_mask |= 1u32 << dim;
-                                    } else if matches!(index.kind, runmat_hir::HirExprKind::End) {
-                                        end_mask |= 1u32 << dim;
-                                    } else {
-                                        // If this index is a Range whose end references End (with or without offset),
-                                        // skip compiling it here; it will be handled by StoreSliceExpr.
-                                        if let Some((_start_end, _step_end, end_expr)) =
-                                            range_dynamic_end_spec(index)
-                                        {
-                                            end_offsets.push((numeric_count, end_expr));
-                                            continue;
-                                        }
-                                        // 1-D end-bounded ranges flow through generic StoreSliceExpr.
-                                        if let Some(off) = end_numeric_expr(index) {
-                                            self.emit(Instr::LoadConst(0.0));
-                                            end_offsets.push((numeric_count, off));
-                                            numeric_count += 1;
-                                            continue;
-                                        }
-                                        self.compile_expr(index)?;
-                                        numeric_count += 1;
-                                    }
-                                }
-                                {
-                                    // Push RHS last so VM pops it first
-                                    // Detect any ranges whose upper bound depends on `end`.
-                                    let mut has_any_range_end = false;
-                                    let mut range_dims: Vec<usize> = Vec::new();
-                                    let mut range_has_step: Vec<bool> = Vec::new();
-                                    let mut range_start_exprs: Vec<Option<EndExpr>> = Vec::new();
-                                    let mut range_step_exprs: Vec<Option<EndExpr>> = Vec::new();
-                                    let mut end_offs: Vec<EndExpr> = Vec::new();
-                                    for (dim, index) in indices.iter().enumerate() {
-                                        if let runmat_hir::HirExprKind::Range(_start, step, _end) =
-                                            &index.kind
-                                        {
-                                            if let Some((start_end, step_end, end_expr)) =
-                                                range_dynamic_end_spec(index)
-                                            {
-                                                has_any_range_end = true;
-                                                range_dims.push(dim);
-                                                range_has_step.push(step.is_some());
-                                                range_start_exprs.push(start_end);
-                                                range_step_exprs.push(step_end);
-                                                end_offs.push(end_expr);
-                                            }
-                                        }
-                                    }
-                                    if has_any_range_end {
-                                        // Push start[, step] for each range in dim order
-                                        for &dim in &range_dims {
-                                            if let runmat_hir::HirExprKind::Range(
-                                                start,
-                                                step,
-                                                _end,
-                                            ) = &indices[dim].kind
-                                            {
-                                                if range_start_exprs[range_dims
-                                                    .iter()
-                                                    .position(|&rd| rd == dim)
-                                                    .unwrap()]
-                                                .is_some()
-                                                {
-                                                    self.emit(Instr::LoadConst(0.0));
-                                                } else {
-                                                    self.compile_expr(start)?;
-                                                }
-                                                if let Some(st) = step {
-                                                    if range_step_exprs[range_dims
-                                                        .iter()
-                                                        .position(|&rd| rd == dim)
-                                                        .unwrap()]
-                                                    .is_some()
-                                                    {
-                                                        self.emit(Instr::LoadConst(0.0));
-                                                    } else {
-                                                        self.compile_expr(st)?;
-                                                    }
-                                                }
-                                            }
-                                        }
-                                        let range_value_count = range_has_step
-                                            .iter()
-                                            .map(|has_step| if *has_step { 2 } else { 1 })
-                                            .sum::<usize>();
-                                        let lvalue_count = 1 + numeric_count + range_value_count;
-                                        let mut lvalue_temps = Vec::with_capacity(lvalue_count);
-                                        for _ in 0..lvalue_count {
-                                            let temp = self.alloc_temp();
-                                            self.emit(Instr::StoreVar(temp));
-                                            lvalue_temps.push(temp);
-                                        }
-                                        lvalue_temps.reverse();
-                                        self.compile_expr(rhs)?;
-                                        let rhs_temp = self.alloc_temp();
-                                        self.emit(Instr::StoreVar(rhs_temp));
-                                        for temp in &lvalue_temps {
-                                            self.emit(Instr::LoadVar(*temp));
-                                        }
-                                        self.emit(Instr::LoadVar(rhs_temp));
-                                        self.emit(Instr::StoreSliceExpr {
-                                            dims: indices.len(),
-                                            numeric_count,
-                                            colon_mask,
-                                            end_mask,
-                                            range_dims,
-                                            range_has_step,
-                                            range_start_exprs,
-                                            range_step_exprs,
-                                            range_end_exprs: end_offs,
-                                            end_numeric_exprs: Vec::new(),
-                                        });
-                                    } else {
-                                        // Attempt packing of function returns or cell expansion for 1-D slices
-                                        let dims_len = indices.len();
-                                        let idx_is_scalar = |e: &HirExpr| -> bool {
-                                            matches!(
-                                                e.kind,
-                                                HirExprKind::Number(_) | HirExprKind::End
-                                            )
-                                        };
-                                        let idx_is_vector = |e: &HirExpr| -> bool {
-                                            matches!(
-                                                e.kind,
-                                                HirExprKind::Colon
-                                                    | HirExprKind::Range(_, _, _)
-                                                    | HirExprKind::Tensor(_)
-                                            )
-                                        };
-                                        let (is_row_slice, is_col_slice) = if dims_len == 2 {
-                                            (
-                                                idx_is_scalar(&indices[0])
-                                                    && idx_is_vector(&indices[1]),
-                                                idx_is_vector(&indices[0])
-                                                    && idx_is_scalar(&indices[1]),
-                                            )
-                                        } else {
-                                            (false, false)
-                                        };
-                                        fn const_vec_len(e: &HirExpr) -> Option<usize> {
-                                            match &e.kind {
-                                                HirExprKind::Number(_) | HirExprKind::End => {
-                                                    Some(1)
-                                                }
-                                                HirExprKind::Tensor(rows) => {
-                                                    Some(rows.iter().map(|r| r.len()).sum())
-                                                }
-                                                HirExprKind::Range(start, step, end) => {
-                                                    if let (
-                                                        HirExprKind::Number(sa),
-                                                        HirExprKind::Number(ea),
-                                                    ) = (&start.kind, &end.kind)
-                                                    {
-                                                        let s: f64 = sa.parse().ok()?;
-                                                        let en: f64 = ea.parse().ok()?;
-                                                        let st: f64 = if let Some(st) = step {
-                                                            if let HirExprKind::Number(x) = &st.kind
-                                                            {
-                                                                x.parse().ok()?
-                                                            } else {
-                                                                return None;
-                                                            }
-                                                        } else {
-                                                            1.0
-                                                        };
-                                                        if st == 0.0 {
-                                                            return None;
-                                                        }
-                                                        let n =
-                                                            ((en - s) / st).floor() as isize + 1;
-                                                        if n <= 0 {
-                                                            Some(0)
-                                                        } else {
-                                                            Some(n as usize)
-                                                        }
-                                                    } else {
-                                                        None
-                                                    }
-                                                }
-                                                HirExprKind::Colon => None,
-                                                _ => None,
-                                            }
-                                        }
-                                        let lvalue_count = 1 + numeric_count;
-                                        let mut lvalue_temps = Vec::with_capacity(lvalue_count);
-                                        for _ in 0..lvalue_count {
-                                            let temp = self.alloc_temp();
-                                            self.emit(Instr::StoreVar(temp));
-                                            lvalue_temps.push(temp);
-                                        }
-                                        lvalue_temps.reverse();
-                                        let mut packed = false;
-                                        if let HirExprKind::FuncCall(fname, fargs) = &rhs.kind {
-                                            if self.functions.contains_key(fname)
-                                                && (dims_len == 1 || is_row_slice || is_col_slice)
-                                            {
-                                                for a in fargs {
-                                                    self.compile_expr(a)?;
-                                                }
-                                                let outc = self
-                                                    .functions
-                                                    .get(fname)
-                                                    .map(|f| f.outputs.len().max(1))
-                                                    .unwrap_or(1);
-                                                self.emit(Instr::CallFunctionMulti(
-                                                    fname.clone(),
-                                                    fargs.len(),
-                                                    outc,
-                                                ));
-                                                self.emit(Instr::Unpack(outc));
-                                                if dims_len == 1 || is_col_slice {
-                                                    self.emit(Instr::PackToCol(outc));
-                                                } else {
-                                                    self.emit(Instr::PackToRow(outc));
-                                                }
-                                                packed = true;
-                                            }
-                                        } else if let HirExprKind::IndexCell(cbase, cidx) =
-                                            &rhs.kind
-                                        {
-                                            // Expand cell into vector matching selected slice length if determinable
-                                            let outc = if dims_len == 1 {
-                                                const_vec_len(&indices[0])
-                                            } else if is_row_slice {
-                                                const_vec_len(&indices[1])
-                                            } else if is_col_slice {
-                                                const_vec_len(&indices[0])
-                                            } else {
-                                                None
-                                            };
-                                            if let Some(n) = outc {
-                                                self.compile_expr(cbase)?;
-                                                // Special case: C{:} => expand all; do not compile colon index
-                                                let expand_all = cidx.len() == 1
-                                                    && matches!(cidx[0].kind, HirExprKind::Colon);
-                                                if expand_all {
-                                                    self.emit(Instr::IndexCellExpand(0, n));
-                                                } else {
-                                                    for i in cidx {
-                                                        self.compile_expr(i)?;
-                                                    }
-                                                    self.emit(Instr::IndexCellExpand(
-                                                        cidx.len(),
-                                                        n,
-                                                    ));
-                                                }
-                                                if dims_len == 1 || is_col_slice {
-                                                    self.emit(Instr::PackToCol(n));
-                                                } else {
-                                                    self.emit(Instr::PackToRow(n));
-                                                }
-                                                packed = true;
-                                            }
-                                        }
-                                        if !packed {
-                                            self.compile_expr(rhs)?;
-                                        }
-                                        let rhs_temp = self.alloc_temp();
-                                        self.emit(Instr::StoreVar(rhs_temp));
-                                        for temp in &lvalue_temps {
-                                            self.emit(Instr::LoadVar(*temp));
-                                        }
-                                        self.emit(Instr::LoadVar(rhs_temp));
-                                        if end_offsets.is_empty() {
-                                            self.emit(Instr::StoreSlice(
-                                                indices.len(),
-                                                numeric_count,
-                                                colon_mask,
-                                                end_mask,
-                                            ));
-                                        } else {
-                                            self.emit(Instr::StoreSliceExpr {
-                                                dims: indices.len(),
-                                                numeric_count,
-                                                colon_mask,
-                                                end_mask,
-                                                range_dims: Vec::new(),
-                                                range_has_step: Vec::new(),
-                                                range_start_exprs: Vec::new(),
-                                                range_step_exprs: Vec::new(),
-                                                range_end_exprs: Vec::new(),
-                                                end_numeric_exprs: end_offsets,
-                                            });
-                                        }
-                                    }
-                                    // Store updated base back to variable
-                                    self.emit(Instr::StoreVar(var_id.0));
-                                }
-                            } else {
-                                // Pure numeric indexing
-                                for index in indices {
-                                    self.compile_expr(index)?;
-                                }
-                                // If RHS is a user function call, request multiple outputs and pack to column for linear targets
-                                if let HirExprKind::FuncCall(fname, fargs) = &rhs.kind {
-                                    if self.functions.contains_key(fname) && indices.len() == 1 {
-                                        for a in fargs {
-                                            self.compile_expr(a)?;
-                                        }
-                                        let outc = self
-                                            .functions
-                                            .get(fname)
-                                            .map(|f| f.outputs.len().max(1))
-                                            .unwrap_or(1);
-                                        self.emit(Instr::CallFunctionMulti(
-                                            fname.clone(),
-                                            fargs.len(),
-                                            outc,
-                                        ));
-                                        self.emit(Instr::Unpack(outc));
-                                        self.emit(Instr::PackToCol(outc));
-                                    } else {
-                                        self.compile_expr(rhs)?;
-                                    }
-                                } else {
-                                    self.compile_expr(rhs)?;
-                                }
-                                self.emit(Instr::StoreIndex(indices.len()));
-                                self.emit(Instr::StoreVar(var_id.0));
-                            }
-                        } else if let runmat_hir::HirExprKind::Member(member_base, field) =
-                            &base.kind
-                        {
-                            // Chain: base is a member access. Evaluate object, load member, update via index, then write member back.
-                            // Evaluate object and get member value
-                            self.compile_member_base_for_assignment(member_base)?;
-                            self.emit(Instr::LoadMemberOrInit(field.clone()));
-                            // Decide slice vs numeric
-                            let has_colon = indices
-                                .iter()
-                                .any(|e| matches!(e.kind, runmat_hir::HirExprKind::Colon));
-                            let has_end = indices
-                                .iter()
-                                .any(|e| matches!(e.kind, runmat_hir::HirExprKind::End));
-                            let has_vector = indices.iter().any(|e| {
-                                matches!(
-                                    e.kind,
-                                    HirExprKind::Range(_, _, _) | HirExprKind::Tensor(_)
-                                ) || matches!(e.ty, runmat_hir::Type::Tensor { .. })
-                            });
-                            if has_colon || has_end || has_vector || indices.len() > 2 {
-                                let mut colon_mask: u32 = 0;
-                                let mut end_mask: u32 = 0;
-                                let mut numeric_count = 0usize;
-                                for (dim, index) in indices.iter().enumerate() {
-                                    if matches!(index.kind, runmat_hir::HirExprKind::Colon) {
-                                        colon_mask |= 1u32 << dim;
-                                    } else if matches!(index.kind, runmat_hir::HirExprKind::End) {
-                                        end_mask |= 1u32 << dim;
-                                    } else {
-                                        self.compile_expr(index)?;
-                                        numeric_count += 1;
-                                    }
-                                }
-                                let lvalue_count = 1 + numeric_count;
-                                let mut lvalue_temps = Vec::with_capacity(lvalue_count);
-                                for _ in 0..lvalue_count {
-                                    let temp = self.alloc_temp();
-                                    self.emit(Instr::StoreVar(temp));
-                                    lvalue_temps.push(temp);
-                                }
-                                lvalue_temps.reverse();
-                                self.compile_expr(rhs)?;
-                                let rhs_temp = self.alloc_temp();
-                                self.emit(Instr::StoreVar(rhs_temp));
-                                for temp in &lvalue_temps {
-                                    self.emit(Instr::LoadVar(*temp));
-                                }
-                                self.emit(Instr::LoadVar(rhs_temp));
-                                self.emit(Instr::StoreSlice(
-                                    indices.len(),
-                                    numeric_count,
-                                    colon_mask,
-                                    end_mask,
-                                ));
-                            } else {
-                                for index in indices {
-                                    self.compile_expr(index)?;
-                                }
-                                self.compile_expr(rhs)?;
-                                self.emit(Instr::StoreIndex(indices.len()));
-                            }
-                            // Now updated member is on stack. Re-evaluate object chain and write back.
-                            self.compile_member_base_for_assignment(member_base)?;
-                            self.emit(Instr::Swap);
-                            self.emit(Instr::StoreMemberOrInit(field.clone()));
-                            self.emit_store_back_member_chain(member_base)?;
-                        } else {
-                            return Err(self.compile_error(
-                                "unsupported lvalue target (index on non-variable/non-member)",
-                            ));
-                        }
-                    }
-                    runmat_hir::HirLValue::IndexCell(base, indices) => {
-                        if let runmat_hir::HirExprKind::Var(var_id) = base.kind {
-                            self.emit(Instr::LoadVar(var_id.0));
-                            for index in indices {
-                                self.compile_expr(index)?;
-                            }
-                            self.compile_expr(rhs)?;
-                            self.emit(Instr::StoreIndexCell(indices.len()));
-                            self.emit(Instr::StoreVar(var_id.0));
-                        } else if let runmat_hir::HirExprKind::Member(member_base, field) =
-                            &base.kind
-                        {
-                            // Load object, load member, perform cell index store, then write member back to object
-                            self.compile_member_base_for_assignment(member_base)?;
-                            self.emit(Instr::LoadMemberOrInit(field.clone()));
-                            for index in indices {
-                                self.compile_expr(index)?;
-                            }
-                            self.compile_expr(rhs)?;
-                            self.emit(Instr::StoreIndexCell(indices.len()));
-                            // Updated member on stack; re-evaluate object chain and write back
-                            self.compile_member_base_for_assignment(member_base)?;
-                            self.emit(Instr::Swap);
-                            self.emit(Instr::StoreMemberOrInit(field.clone()));
-                            self.emit_store_back_member_chain(member_base)?;
-                        } else {
-                            // Fallback: evaluate base, indices, rhs, and store (for object chains via subsasgn)
-                            self.compile_expr(base)?;
-                            for index in indices {
-                                self.compile_expr(index)?;
-                            }
-                            self.compile_expr(rhs)?;
-                            self.emit(Instr::StoreIndexCell(indices.len()));
-                        }
-                    }
-                    runmat_hir::HirLValue::Member(base, field) => {
-                        self.compile_member_base_for_assignment(base)?;
-                        self.compile_expr(rhs)?;
-                        if Self::allow_implicit_struct_materialization(&base.ty) {
-                            self.emit(Instr::StoreMemberOrInit(field.clone()));
-                        } else {
-                            self.emit(Instr::StoreMember(field.clone()));
-                        }
-                        self.emit_store_back_member_chain(base)?;
-                    }
-                    runmat_hir::HirLValue::MemberDynamic(base, name_expr) => {
-                        self.compile_member_base_for_assignment(base)?;
-                        self.compile_expr(name_expr)?;
-                        self.compile_expr(rhs)?;
-                        if Self::allow_implicit_struct_materialization(&base.ty) {
-                            self.emit(Instr::StoreMemberDynamicOrInit);
-                        } else {
-                            self.emit(Instr::StoreMemberDynamic);
-                        }
-                        self.emit_store_back_member_chain(base)?;
-                    }
-                    _ => return Err(self.compile_error("unsupported lvalue target")),
-                }
-            }
+            HirStmt::AssignLValue(lv, rhs, _, _) => self.compile_assign_lvalue(lv, rhs)?,
             HirStmt::Global(vars, _) => {
                 let ids: Vec<usize> = vars.iter().map(|(v, _n)| v.0).collect();
                 let names: Vec<String> = vars.iter().map(|(_v, n)| n.clone()).collect();
@@ -2871,63 +1944,7 @@ impl Compiler {
                     self.emit(Instr::CallMethod(method.clone(), args.len()));
                 }
             },
-            HirExprKind::AnonFunc { params, body } => {
-                // Collect free variables in body (in order of first appearance)
-                use std::collections::{HashMap, HashSet};
-                let mut seen: HashSet<runmat_hir::VarId> = HashSet::new();
-                let mut captures_order: Vec<runmat_hir::VarId> = Vec::new();
-                let bound: HashSet<runmat_hir::VarId> = params.iter().cloned().collect();
-                self.collect_free_vars(body, &bound, &mut seen, &mut captures_order);
-
-                // Build placeholder VarIds for captures and parameters
-                let capture_count = captures_order.len();
-                let mut placeholder_params: Vec<runmat_hir::VarId> =
-                    Vec::with_capacity(capture_count + params.len());
-                for i in 0..capture_count {
-                    placeholder_params.push(runmat_hir::VarId(i));
-                }
-                for j in 0..params.len() {
-                    placeholder_params.push(runmat_hir::VarId(capture_count + j));
-                }
-                let output_id = runmat_hir::VarId(capture_count + params.len());
-
-                // Remap body vars: free vars -> capture placeholders; param vars -> shifted placeholders
-                let mut var_map: HashMap<runmat_hir::VarId, runmat_hir::VarId> = HashMap::new();
-                for (i, old) in captures_order.iter().enumerate() {
-                    var_map.insert(*old, runmat_hir::VarId(i));
-                }
-                for (j, old) in params.iter().enumerate() {
-                    var_map.insert(*old, runmat_hir::VarId(capture_count + j));
-                }
-                let remapped_body = runmat_hir::remapping::remap_expr(body, &var_map);
-                let func_body = vec![runmat_hir::HirStmt::Assign(
-                    output_id,
-                    remapped_body,
-                    true,
-                    runmat_hir::Span::default(),
-                )];
-
-                // Synthesize function name and register
-                let synthesized = format!("__anon_{}", self.functions.len());
-                let user_func = UserFunction {
-                    name: synthesized.clone(),
-                    params: placeholder_params,
-                    outputs: vec![output_id],
-                    body: func_body,
-                    local_var_count: capture_count + params.len() + 1,
-                    has_varargin: false,
-                    has_varargout: false,
-                    var_types: vec![Type::Unknown; capture_count + params.len() + 1],
-                    source_id: None,
-                };
-                self.functions.insert(synthesized.clone(), user_func);
-
-                // Emit capture values on stack then create closure
-                for old in &captures_order {
-                    self.emit(Instr::LoadVar(old.0));
-                }
-                self.emit(Instr::CreateClosure(synthesized, capture_count));
-            }
+            HirExprKind::AnonFunc { params, body } => self.compile_anon_func(params, body)?,
             HirExprKind::FuncHandle(name) => {
                 self.emit(Instr::LoadString(name.clone()));
                 self.emit(Instr::CallBuiltin("make_handle".to_string(), 1));
